+++ src/Native/Builtin.js
@@ -1,8 +1,10 @@
 var _coowry$elm_js_standard_builtin$Native_Builtin = function(elm) {
 
-    var now = Date.now();
+    var now = function() { return Date.now(); };
+
+    var newDate = function(){ return new Date(); };
     
     return  { now : now,
-              newDate: new Date()
+              newDate: newDate
             };
 }();