+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ## Next release
 
+- [F] Builtin.Date.now and Builtin.Date.new are functions now, to
+  allow each execution to return the current date.
+
 ## 1.0.1
 
 - [F] fixing the native prefix from aherranz to coowry.